Refactor by individuating static classes and initial model specific code.
